+++ flowmachine/flowmachine/features/utilities/event_table_subset.py
@@ -3,16 +3,16 @@
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 
-import datetime
-import pandas as pd
 import warnings
 from sqlalchemy import select
 from typing import List, Optional, Tuple
 
-from ...core import Query, Table
-from ...core.context import get_db
-from ...core.errors import MissingDateError
-from ...core.sqlalchemy_utils import (
+from flowmachine.core.query import Query
+from flowmachine.core.context import get_db
+from flowmachine.core.errors import MissingDateError
+from flowmachine.core.events_table import events_table_map
+from flowmachine.core.preflight import pre_flight
+from flowmachine.core.sqlalchemy_utils import (
     get_sqlalchemy_table_definition,
     make_sqlalchemy_column_from_flowmachine_column_description,
     get_sql_string,
@@ -32,19 +32,15 @@ class EventTableSubset(Query):
 
     Parameters
     ----------
-    start : str, default None
-        iso format date range for the beginning of the time frame, e.g.
-        2016-01-01 or 2016-01-01 14:03:01. If None, it will use the
-        earliest date seen in the `events.calls` table.
-    stop : str, default None
-        As above. If None, it will use the latest date seen in the
-        `events.calls` table.
-    hours : tuple of ints, default None
+    start, stop : str
+        iso format date range for the beginning and end of the time frame, e.g.
+        2016-01-01 or 2016-01-01 14:03:01.
+    hours : tuple of ints, default 'None'
         Subset the result within certain hours, e.g. (4,17)
         This will subset the query only with these hours, but
         across all specified days. Or set to 'all' to include
         all hours.
-    table : str, default 'events.calls'
+    table : str, default 'calls'
         schema qualified name of the table which the analysis is
         based upon
     subscriber_identifier : {'msisdn', 'imei'}, default 'msisdn'
@@ -72,13 +68,13 @@ class EventTableSubset(Query):
     def __init__(
         self,
         *,
-        start=None,
-        stop=None,
+        start,
+        stop,
         hours: Optional[Tuple[int, int]] = None,
         hour_slices=None,
-        table="events.calls",
+        table="calls",
         subscriber_subset=None,
-        columns=["*"],
+        columns=None,
         subscriber_identifier="msisdn",
     ):
         if hours == "all":
@@ -121,12 +117,9 @@ class EventTableSubset(Query):
         self.hours = hours
         self.subscriber_subsetter = make_subscriber_subsetter(subscriber_subset)
         self.subscriber_identifier = subscriber_identifier.lower()
-        if columns == ["*"]:
-            self.table_ORIG = Table(table)
-            columns = self.table_ORIG.column_names
-        else:
-            self.table_ORIG = Table(table, columns=columns)
-        self.columns = set(columns)
+        self.table_ORIG = events_table_map[table](columns=columns)
+
+        self.columns = set(self.table_ORIG.column_names)
         try:
             self.columns.remove(subscriber_identifier)
             self.columns.add(f"{subscriber_identifier} AS subscriber")
@@ -140,57 +133,31 @@ class EventTableSubset(Query):
                 )
         self.columns = sorted(self.columns)
 
-        self.sqlalchemy_table = get_sqlalchemy_table_definition(
-            self.table_ORIG.fully_qualified_table_name,
-            engine=get_db().engine,
-        )
-
         if self.start == self.stop:
             raise ValueError("Start and stop are the same.")
 
         super().__init__()
 
-        # This needs to happen after the parent classes init method has been
-        # called as it relies upon the connection object existing
-        self._check_dates()
-
     @property
     def column_names(self) -> List[str]:
         return [c.split(" AS ")[-1] for c in self.columns]
 
-    def _check_dates(self):
+    @pre_flight
+    def check_dates(self):
+        logger.debug("Checking dates are valid.")
         # Handle the logic for dealing with missing dates.
         # If there are no dates present, then we raise an error
         # if some are present, but some are missing we raise a
         # warning.
-        # If the subscriber does not pass a start or stop date, then we take
-        # the min/max date in the events.calls table
-        if self.start is None:
-            d1 = (
-                get_db()
-                .min_date(self.table_ORIG.fully_qualified_table_name.split(".")[1])
-                .strftime("%Y-%m-%d %H:%M:%S")
-            )
-        else:
-            d1 = self.start.split()[0]
-
-        if self.stop is None:
-            d2 = (
-                get_db()
-                .max_date(self.table_ORIG.fully_qualified_table_name.split(".")[1])
-                .strftime("%Y-%m-%d %H:%M:%S")
-            )
-        else:
-            d2 = self.stop.split()[0]
+        d1 = self.start.split()[0]
+        d2 = self.stop.split()[0]
 
         all_dates = list_of_dates(d1, d2)
         # Slightly annoying feature, but if the subscriber passes a date such as '2016-01-02'
         # this will be interpreted as midnight, so we don't want to include this in our
         # calculations. Check for this here, an if this is the case pop the final element
         # of the list
-        if (self.stop is not None) and (
-            len(self.stop) == 10 or self.stop.endswith("00:00:00")
-        ):
+        if len(self.stop) == 10 or self.stop.endswith("00:00:00"):
             all_dates.pop(-1)
         # This will be a true false list for whether each of the dates
         # is present in the database
@@ -216,34 +183,32 @@ class EventTableSubset(Query):
                 stacklevel=2,
             )
 
-    def _make_query_with_sqlalchemy(self):
+    def _make_query(self):
+        sqlalchemy_table = get_sqlalchemy_table_definition(
+            self.table_ORIG.fully_qualified_table_name,
+            engine=get_db().engine,
+        )
         sqlalchemy_columns = [
             make_sqlalchemy_column_from_flowmachine_column_description(
-                self.sqlalchemy_table, column_str
+                sqlalchemy_table, column_str
             )
             for column_str in self.columns
         ]
         select_stmt = select(*sqlalchemy_columns)
 
         if self.start is not None:
-            select_stmt = select_stmt.where(
-                self.sqlalchemy_table.c.datetime >= self.start
-            )
+            select_stmt = select_stmt.where(sqlalchemy_table.c.datetime >= self.start)
         if self.stop is not None:
-            select_stmt = select_stmt.where(
-                self.sqlalchemy_table.c.datetime < self.stop
-            )
+            select_stmt = select_stmt.where(sqlalchemy_table.c.datetime < self.stop)
 
         select_stmt = select_stmt.where(
-            self.hour_slices.get_subsetting_condition(self.sqlalchemy_table.c.datetime)
+            self.hour_slices.get_subsetting_condition(sqlalchemy_table.c.datetime)
         )
         select_stmt = self.subscriber_subsetter.apply_subset_if_needed(
             select_stmt, subscriber_identifier=self.subscriber_identifier
         )
 
         return get_sql_string(select_stmt)
-
-    _make_query = _make_query_with_sqlalchemy
 
     @property
     def fully_qualified_table_name(self):