+++ flowmachine/tests/test_daily_location.py
@@ -6,6 +6,7 @@ import pytest
 
 from flowmachine.core.errors import MissingDateError
 from flowmachine.core import make_spatial_unit
+from flowmachine.core.errors.flowmachine_errors import PreFlightFailedException
 from flowmachine.features import daily_location, MostFrequentLocation
 
 
@@ -77,5 +78,13 @@ def test_daily_locs_errors():
     daily_location() errors when we ask for a date that does not exist.
     """
 
+    with pytest.raises(PreFlightFailedException) as exc:
+        daily_location("2016-01-31").preflight()
     with pytest.raises(MissingDateError):
-        daily_location("2016-01-31")
+        raise exc.value.errors[
+            "<Query of type: EventTableSubset, query_id: '607268fcdebd1e536d30256ee0ee40e1'>"
+        ][0]
+    with pytest.raises(MissingDateError):
+        raise exc.value.errors[
+            "<Query of type: EventTableSubset, query_id: '8533ca0424a0030218ec75f691c6d465'>"
+        ][0]