+++ flowmachine/tests/test_redacted_total_events.py
@@ -23,13 +23,13 @@ def test_all_above_threshold(get_dataframe):
             "2016-01-02",
             spatial_unit=make_spatial_unit("cell"),
             interval="day",
-            table=["events.calls"],
+            table=["calls"],
         )
     )
     us = get_dataframe(
         RedactedUniqueSubscriberCounts(
             unique_subscriber_counts=UniqueSubscriberCounts(
-                "2016-01-01", "2016-01-02", table=["events.calls"]
+                "2016-01-01", "2016-01-02", table=["calls"]
             )
         )
     )
@@ -49,7 +49,7 @@ def test_all_above_threshold_hour_bucket(get_dataframe):
             "2016-01-02",
             spatial_unit=make_spatial_unit("cell"),
             interval="hour",
-            table=["events.calls"],
+            table=["calls"],
         )
     )
 
@@ -67,7 +67,7 @@ def test_all_above_threshold_minute_bucket(get_dataframe):
             "2016-01-01 13:00",
             spatial_unit=make_spatial_unit("cell"),
             interval="min",
-            table=["events.calls"],
+            table=["calls"],
         )
     )